+++ ocr/runtime/ocr-x86/src/policy-domain/hc/hc-policy.h
@@ -28,7 +28,7 @@ typedef struct {
 
 typedef struct {
     ocrPolicyDomain_t base;
-    u64 (*adaptObjectiveFct)(u64);
+    u64 (*adaptObjectiveFct)(int);
 } ocrPolicyDomainHc_t;
 
 ocrPolicyDomainFactory_t *newPolicyDomainFactoryHc(ocrParamList_t *perType);
